ai_service/image_analyzer.py
import hashlib
import io
import random
from typing import Dict, Any, Optional

try:
    from PIL import Image
    HAS_PIL = True
except ImportError:
    HAS_PIL = False

try:
    import numpy as np
    HAS_NUMPY = True
except ImportError:
    HAS_NUMPY = False

from disease_database import DISEASE_DATABASE

class ImageAnalyzer:
    def __init__(self):
        self.disease_db = DISEASE_DATABASE
        self._seed_counter = 0
        
    def generate_image_hash(self, image_data: bytes) -> str:
        if HAS_PIL and HAS_NUMPY:
            try:
                return self._perceptual_hash(image_data)
            except Exception:
                pass
        
        return hashlib.sha256(image_data).hexdigest()
    
    def _perceptual_hash(self, image_data: bytes) -> str:
        image = Image.open(io.BytesIO(image_data))
        
        image = image.convert('L').resize((8, 8), Image.Resampling.LANCZOS)
        
        pixels = list(image.getdata())
        
        avg = sum(pixels) / len(pixels)
        
        bits = ''.join(['1' if p > avg else '0' for p in pixels])
        
        hex_hash = '{:016x}'.format(int(bits, 2))
        
        return hex_hash
    
    def _get_crop_category(self, crop_type: str) -> str:
        crop_type = crop_type.lower()
        
        if any(c in crop_type for c in ['rice', '水稻', '稻']):
            return 'rice'
        elif any(c in crop_type for c in ['vegetable', '蔬菜', '番茄', '黄瓜', '白菜', '茄子', '辣椒']):
            return 'vegetable'
        elif any(c in crop_type for c in ['fruit', '果树', '苹果', '柑橘', '梨', '桃']):
            return 'fruit_tree'
        
        return 'rice'
    
    def analyze_image(self, image_data: bytes, image_name: str, crop_type: str) -> Dict[str, Any]:
        image_hash = self.generate_image_hash(image_data)
        
        crop_category = self._get_crop_category(crop_type)
        
        diseases = self.disease_db.get(crop_category, [])
        
        if not diseases:
            diseases = self.disease_db.get('rice', [])
        
        selected_disease = self._select_disease_by_hash(image_hash, diseases, crop_type)
        
        confidence = self._calculate_confidence(image_hash, selected_disease)
        
        similar_cases_json = self._format_similar_cases(selected_disease.get('similar_cases', []))
        
        result = {
            "success": True,
            "disease_name": selected_disease.get('name', '未知病害'),
            "disease_type": selected_disease.get('type', '未知类型'),
            "confidence": confidence,
            "symptoms": selected_disease.get('symptoms', ''),
            "causes": selected_disease.get('causes', ''),
            "recommended_actions": '；'.join(selected_disease.get('recommended_actions', [])),
            "severity": self._estimate_severity(image_hash, confidence),
            "similar_cases": similar_cases_json,
            "image_hash": image_hash,
            "crop_category": crop_category
        }
        
        return result
    
    def _select_disease_by_hash(self, image_hash: str, diseases: list, crop_type: str) -> Dict:
        if not diseases:
            return {}
        
        hash_int = int(image_hash[:8], 16)
        
        index = hash_int % len(diseases)
        
        self._seed_counter = (self._seed_counter + 1) % 100
        
        return diseases[index]
    
    def _calculate_confidence(self, image_hash: str, disease: Dict) -> float:
        hash_sum = sum(int(c, 16) for c in image_hash[:16])
        
        base_confidence = 0.7 + (hash_sum % 30) / 100
        
        disease_name = disease.get('name', '')
        if '稻瘟病' in disease_name:
            base_confidence = min(0.95, base_confidence + 0.05)
        elif '纹枯病' in disease_name:
            base_confidence = min(0.92, base_confidence + 0.02)
        
        return round(base_confidence, 2)
    
    def _estimate_severity(self, image_hash: str, confidence: float) -> str:
        hash_char = image_hash[0]
        
        severity_map = {
            '0': '轻度', '1': '轻度', '2': '轻度', '3': '轻度',
            '4': '中度', '5': '中度', '6': '中度', '7': '中度',
            '8': '重度', '9': '重度', 'a': '重度', 'b': '重度',
            'c': '轻度', 'd': '中度', 'e': '中度', 'f': '重度'
        }
        
        severity = severity_map.get(hash_char.lower(), '中度')
        
        if confidence > 0.9:
            severity = '重度' if hash_char in ['8', '9', 'e', 'f'] else severity
        elif confidence < 0.75:
            severity = '轻度'
        
        return severity
    
    def _format_similar_cases(self, cases: list) -> str:
        if not cases:
            return '[]'
        
        import json
        return json.dumps(cases, ensure_ascii=False)
    
    def get_disease_details(self, disease_name: str, crop_category: str = 'rice') -> Optional[Dict]:
        diseases = self.disease_db.get(crop_category, [])
        
        for disease in diseases:
            if disease.get('name') == disease_name:
                return disease
        
        for category, diseases in self.disease_db.items():
            for disease in diseases:
                if disease.get('name') == disease_name:
                    return disease
        
        return None
    
    def generate_treatment_plan(self, disease_name: str, severity: str, 
                                  crop_category: str = 'rice') -> Dict[str, Any]:
        disease = self.get_disease_details(disease_name, crop_category)
        
        if not disease:
            return {
                "success": False,
                "message": "未找到该病害信息"
            }
        
        treatment_plans = disease.get('treatment_plan', {})
        medications = disease.get('medications', [])
        prevention_tips = disease.get('prevention_tips', [])
        
        plan = treatment_plans.get(severity, treatment_plans.get('中度', ''))
        
        return {
            "success": True,
            "disease_name": disease_name,
            "severity": severity,
            "treatment_plan": plan,
            "medications": medications,
            "prevention_tips": prevention_tips,
            "recommended_actions": disease.get('recommended_actions', [])
        }
