ai_service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
numpy==1.26.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
pandas==2.1.3
python-dotenv==1.0.0
werkzeug==3.0.1
